app/requirements.txt
Flask
mysql-connector-python
configparser
